LPS-183342 migrate the logic of getPart() and getParts() to ClientDataRequestHelper
