+++ portal-impl/src/com/liferay/portlet/internal/ClientDataRequestImpl.java
@@ -14,25 +14,18 @@
 
 package com.liferay.portlet.internal;
 
-import com.liferay.portal.kernel.model.Portlet;
+import com.liferay.portal.kernel.portlet.ClientDataRequestHelperUtil;
 import com.liferay.portal.kernel.servlet.HttpMethods;
-import com.liferay.portal.kernel.upload.FileItem;
-import com.liferay.portal.kernel.upload.UploadRequest;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
 
 import java.io.BufferedReader;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UnsupportedEncodingException;
 
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
@@ -40,7 +33,6 @@ import javax.portlet.ClientDataRequest;
 import javax.portlet.PortletException;
 import javax.portlet.RenderParameters;
 
-import javax.servlet.ServletRequestWrapper;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.Part;
 
@@ -78,55 +70,14 @@ public abstract class ClientDataRequestImpl
 
 	@Override
 	public Part getPart(String name) throws IOException, PortletException {
-		UploadRequest uploadRequest = _getUploadRequest(
-			getHttpServletRequest());
-
-		if (uploadRequest == null) {
-			return null;
-		}
-
-		Map<String, FileItem[]> multipartParameterMap =
-			uploadRequest.getMultipartParameterMap();
-
-		FileItem[] fileItems = multipartParameterMap.get(name);
-
-		if ((fileItems == null) || (fileItems.length == 0)) {
-			return null;
-		}
-
-		Portlet portlet = getPortlet();
-
-		return new PartImpl(fileItems[0], portlet.getMultipartLocation());
+		return ClientDataRequestHelperUtil.getPart(
+			name, getHttpServletRequest(), getPortlet());
 	}
 
 	@Override
 	public Collection<Part> getParts() throws IOException, PortletException {
-		UploadRequest uploadRequest = _getUploadRequest(
-			getHttpServletRequest());
-
-		if (uploadRequest == null) {
-			return Collections.emptySet();
-		}
-
-		List<Part> parts = new ArrayList<>();
-
-		Map<String, FileItem[]> multipartParameterMap =
-			uploadRequest.getMultipartParameterMap();
-
-		Portlet portlet = getPortlet();
-
-		for (Map.Entry<String, FileItem[]> entry :
-				multipartParameterMap.entrySet()) {
-
-			FileItem[] fileItems = entry.getValue();
-
-			for (FileItem fileItem : fileItems) {
-				parts.add(
-					new PartImpl(fileItem, portlet.getMultipartLocation()));
-			}
-		}
-
-		return parts;
+		return ClientDataRequestHelperUtil.getParts(
+			getHttpServletRequest(), getPortlet());
 	}
 
 	@Override
@@ -208,107 +159,6 @@ public abstract class ClientDataRequestImpl
 		}
 	}
 
-	private UploadRequest _getUploadRequest(Object request) {
-		while (true) {
-			if (request instanceof UploadRequest) {
-				return (UploadRequest)request;
-			}
-
-			if (request instanceof ServletRequestWrapper) {
-				ServletRequestWrapper servletRequestWrapper =
-					(ServletRequestWrapper)request;
-
-				request = servletRequestWrapper.getRequest();
-			}
-			else {
-				return null;
-			}
-		}
-	}
-
 	private boolean _calledGetReader;
-
-	private static final class PartImpl implements Part {
-
-		@Override
-		public void delete() throws IOException {
-			_fileItem.delete();
-		}
-
-		@Override
-		public String getContentType() {
-			return _fileItem.getContentType();
-		}
-
-		@Override
-		public String getHeader(String name) {
-			return _fileItem.getHeader(name);
-		}
-
-		@Override
-		public Collection<String> getHeaderNames() {
-			return _fileItem.getHeaderNames();
-		}
-
-		@Override
-		public Collection<String> getHeaders(String name) {
-			return _fileItem.getHeaders(name);
-		}
-
-		@Override
-		public InputStream getInputStream() throws IOException {
-			return _fileItem.getInputStream();
-		}
-
-		@Override
-		public String getName() {
-			return _fileItem.getFieldName();
-		}
-
-		@Override
-		public long getSize() {
-			return _fileItem.getSize();
-		}
-
-		@Override
-		public String getSubmittedFileName() {
-			return _fileItem.getFileName();
-		}
-
-		@Override
-		public void write(String fileName) throws IOException {
-			if (Validator.isNull(fileName)) {
-				throw new IOException("Invalid file name");
-			}
-
-			try {
-				File file = new File(fileName);
-
-				if (!file.isAbsolute() &&
-					Validator.isNotNull(_multipartLocation)) {
-
-					File multipartLocation = new File(_multipartLocation);
-
-					if (multipartLocation.isDirectory()) {
-						file = new File(multipartLocation, fileName);
-					}
-				}
-
-				_fileItem.write(file);
-			}
-			catch (Exception exception) {
-				throw new IOException(exception);
-			}
-		}
-
-		private PartImpl(FileItem fileItem, String multipartLocation) {
-			_fileItem = fileItem;
-			_multipartLocation = multipartLocation;
-		}
-
-		private final FileItem _fileItem;
-		private final String _multipartLocation;
-
-	}
 
 }
